make code more amenable to compiling with mingw for cross-compiling

+++ OpenEXR/IlmImf/ImfOptimizedPixelReading.h
@@ -70,7 +70,7 @@ EXR_FORCEINLINE
 bool
 isPointerSSEAligned (const void* EXR_RESTRICT pPointer)
 {
-    unsigned long trailingBits = ((unsigned long)pPointer) & 15;
+    uintptr_t trailingBits = ((uintptr_t)pPointer) & 15;
     return trailingBits == 0;
 }
